split posts into interactions and post
